requirements.txt
Flask==1.1.2
PyYAML==5.3.1